Fix Dependabot logic for SNAPSHOTs
The Dependabot has a logic which we find as wrong.
It updates from the current `M/RC/SNAPSHOT` not to the GA of that version, but to the next version's `SNAPSHOT`

* Fix `spring-merge-dependabot-pr.yml` to compare `new-version` and `previous-version` metadata without suffixes.
For example, compare the current ` 1.0.0-RC1 ` with a new `1.0.1-SNAPSHOT`.
This means there is a GA we would expect an update to, but somehow Dependabot skips that for us.
* Close its PRs in this case.
Only allow update to a new `SNAPSHOT` if `new-version` and `previous-version` version bases are equal